Installed react, redux framework frontend

+++ server/controllers/listController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler');
 
 const List = require('../models/listModel');
+const User = require('../models/userModel');
 
-// @Desc Gets the lists stored
+// @Desc Gets a specific user's lists
 // @Route GET /api/lists
 // @Access Private
 const getLists = asyncHandler(async (req, res) => {
-    const lists = await List.find();
+    const lists = await List.find({ user: req.user.id });
 
     res.status(200).json(lists);
 });
@@ -21,7 +22,8 @@ const setList = asyncHandler(async (req, res) => {
     }
 
     const list = await List.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id,
     });
 
     res.status(200).json(list);
@@ -36,6 +38,22 @@ const updateList = asyncHandler(async (req, res) => {
     if (!list) {
         res.status(400);
         throw new Error('List not found');
+    }
+
+    // user will just be the ID of the user
+    const user = await User.findById(req.user.id);
+
+    // Checks for user, if an ID was found
+    if (!user) {
+        res.status(401);
+        throw new Error('User not found!');
+    }
+
+    // if the list's user id is not equal to the user that is trying to view the list
+    // Make sure logged in user matches the same as the one that created the list
+    if (list.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User not authorized');
     }
 
     const updatedList = await List.findByIdAndUpdate(req.params.id, req.body,
@@ -55,6 +73,22 @@ const deleteList = asyncHandler(async (req, res) => {
     if (!list) {
         res.status(400);
         throw new Error('List not found');
+    }
+
+    // user will just be the ID of the user
+    const user = await User.findById(req.user.id);
+
+    // Checks for user, if an ID was found
+    if (!user) {
+        res.status(401);
+        throw new Error('User not found!');
+    }
+
+    // if the list's user id is not equal to the user that is trying to view the list
+    // Make sure logged in user matches the same as the one that created the list
+    if (list.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User no authorized');
     }
 
     await list.deleteOne()
